chore: fix forge lint warnings and add health tools

+++ contracts/script/Deploy.s.sol
@@ -10,7 +10,7 @@ contract Deploy is BaseScript {
         address admin = vm.envOr({ name: "ORACLE_ADMIN", defaultValue: broadcaster });
         address indexer = vm.envOr({ name: "ORACLE_INDEXER", defaultValue: broadcaster });
         uint256 periodDuration = vm.envOr({ name: "ORACLE_PERIOD", defaultValue: uint256(1 hours) });
-        uint32 maxSamples = vm.envOr({ name: "ORACLE_MAX_SAMPLES", defaultValue: uint32(256) });
+        uint32 maxSamples = uint32(vm.envOr({ name: "ORACLE_MAX_SAMPLES", defaultValue: uint256(256) }));
         uint256 submissionScale = vm.envOr({ name: "ORACLE_SUBMISSION_SCALE", defaultValue: uint256(1e4) });
 
         oracle = new ZKOracle(indexer, periodDuration, maxSamples, submissionScale, admin);
